Add files via upload
Here we will find a .au3 file (zerofileupload.au3) which is Auto IT script. Inside the script we have written that AutoIT will provide th file path link into the Windows Explorer and upload the file in OneDrive. Then we have compiled the .au3 file and created a .exe file. Then we have opened the .exe file (zerofileupload.exe) in Selenium which will complete the file uploading process.