Add files via upload
These are files which will help to upload a file into OneDrive. The .au3 file has script within it. The script will input the file path into the File Upload pop-up and upload the file in OneDrive. By compiling the .au3 file, we got the .exe file. We will execute the .exe file in Selenium. Thus, we will be able to upload a file with contents in OneDrive through Selenium.